* Restored validation in local files * made code compilable
